+++ tests/policy_definition/update_management/maintenance_resources_assignment/locals.tf
@@ -1,0 +1,4 @@
+locals {
+  role_definition_ids      = var.role_definition_ids
+  regex_pattern_id_to_name = "[\\w-]+$"
+}