+++ tests/policy_definition/update_management/maintenance_resources_assignment/outputs.tf
@@ -1,0 +1,4 @@
+output "policy_assignment" {
+  description = "The policy assignment created."
+  value       = azurerm_subscription_policy_assignment.this
+}