+++ tests/policy_definition/update_management/maintenance_resources_compliance/terraform.tf
@@ -1,0 +1,11 @@
+terraform {
+  required_providers {
+    azapi = {
+      source = "azure/azapi"
+    }
+    azurerm = {
+      source = "hashicorp/azurerm"
+    }
+  }
+  required_version = ">=1.6.0"
+}