+++ tests/policy_definition/update_management/maintenance_resources_setup/terraform.tf
@@ -1,0 +1,12 @@
+terraform {
+  required_providers {
+    azurerm = {
+      source                = "hashicorp/azurerm"
+      configuration_aliases = [azurerm.management]
+    }
+    random = {
+      source = "hashicorp/random"
+    }
+  }
+  required_version = ">=1.6.0"
+}