+++ tests/policy_definition/update_management/maintenance_resources_vm/terraform.tf
@@ -1,0 +1,9 @@
+terraform {
+  required_providers {
+    azurerm = {
+      source                = "hashicorp/azurerm"
+      configuration_aliases = [azurerm.management]
+    }
+  }
+  required_version = ">=1.6.0"
+}