+++ CHANGELOG.md
@@ -5,14 +5,20 @@ All notable changes to this module will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this module adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
 ## [Unreleased]
 
+
 ### Added
+
 - Initiative deploying new Azure Monitoring Agent and associating data collection rules
+- added policy set for Azure Update Manager
 - Added require tags policy set defitions and assignment
 - Policy definition to allow vm/vmss SKUs
 
+
 ### Removed
+
 - Policies deploying the old Microsoft Monitoring Agent
 
 ## [2.3.0] - 2024-01-12